requirements.txt
Flask
Flask-Compress
numpy
opencv-python
Pillow
ImageHash
google-auth
requests
python-dotenv
ultralytics
